LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/animation/rotate-composition-expected.txt


FAIL Compositing: property <rotate> underlying [100deg] from add [10deg] to add [30deg] at (-1) should be [90deg] assert_equals: expected "90deg " but got "100deg "
FAIL Compositing: property <rotate> underlying [100deg] from add [10deg] to add [30deg] at (0) should be [110deg] assert_equals: expected "110deg " but got "100deg "
FAIL Compositing: property <rotate> underlying [100deg] from add [10deg] to add [30deg] at (0.25) should be [115deg] assert_equals: expected "115deg " but got "100deg "
FAIL Compositing: property <rotate> underlying [100deg] from add [10deg] to add [30deg] at (0.75) should be [125deg] assert_equals: expected "125deg " but got "100deg "
FAIL Compositing: property <rotate> underlying [100deg] from add [10deg] to add [30deg] at (1) should be [130deg] assert_equals: expected "130deg " but got "100deg "
FAIL Compositing: property <rotate> underlying [100deg] from add [10deg] to add [30deg] at (2) should be [150deg] assert_equals: expected "150deg " but got "100deg "
FAIL Compositing: property <rotate> underlying [1 0 0 200deg] from add [1 0 0 -100deg] to replace [1 0 0 40deg] at (-1) should be [1 0 0 160deg] assert_equals: expected "1 0 0 160deg " but got "1 0 0 200deg "
FAIL Compositing: property <rotate> underlying [1 0 0 200deg] from add [1 0 0 -100deg] to replace [1 0 0 40deg] at (0) should be [1 0 0 100deg] assert_equals: expected "1 0 0 100deg " but got "1 0 0 200deg "
FAIL Compositing: property <rotate> underlying [1 0 0 200deg] from add [1 0 0 -100deg] to replace [1 0 0 40deg] at (0.25) should be [1 0 0 85deg] assert_equals: expected "1 0 0 85deg " but got "1 0 0 200deg "
FAIL Compositing: property <rotate> underlying [1 0 0 200deg] from add [1 0 0 -100deg] to replace [1 0 0 40deg] at (0.75) should be [1 0 0 55deg] assert_equals: expected "1 0 0 55deg " but got "1 0 0 200deg "
FAIL Compositing: property <rotate> underlying [1 0 0 200deg] from add [1 0 0 -100deg] to replace [1 0 0 40deg] at (1) should be [1 0 0 40deg] assert_equals: expected "1 0 0 40deg " but got "1 0 0 200deg "
FAIL Compositing: property <rotate> underlying [1 0 0 200deg] from add [1 0 0 -100deg] to replace [1 0 0 40deg] at (2) should be [1 0 0 -20deg] assert_equals: expected "1 0 0 - 20deg " but got "1 0 0 200deg "
FAIL Compositing: property <rotate> underlying [0 1 0 -40deg] from replace [0 1 0 50deg] to add [0 1 0 10deg] at (-1) should be [0 1 0 130deg] assert_equals: expected "0 1 0 130deg " but got "0 1 0 - 40deg "
FAIL Compositing: property <rotate> underlying [0 1 0 -40deg] from replace [0 1 0 50deg] to add [0 1 0 10deg] at (0) should be [0 1 0 50deg] assert_equals: expected "0 1 0 50deg " but got "0 1 0 - 40deg "
FAIL Compositing: property <rotate> underlying [0 1 0 -40deg] from replace [0 1 0 50deg] to add [0 1 0 10deg] at (0.25) should be [0 1 0 30deg] assert_equals: expected "0 1 0 30deg " but got "0 1 0 - 40deg "
FAIL Compositing: property <rotate> underlying [0 1 0 -40deg] from replace [0 1 0 50deg] to add [0 1 0 10deg] at (0.75) should be [0 1 0 -10deg] assert_equals: expected "0 1 0 - 10deg " but got "0 1 0 - 40deg "
FAIL Compositing: property <rotate> underlying [0 1 0 -40deg] from replace [0 1 0 50deg] to add [0 1 0 10deg] at (1) should be [0 1 0 -30deg] assert_equals: expected "0 1 0 - 30deg " but got "0 1 0 - 40deg "
FAIL Compositing: property <rotate> underlying [0 1 0 -40deg] from replace [0 1 0 50deg] to add [0 1 0 10deg] at (2) should be [0 1 0 -110deg] assert_equals: expected "0 1 0 - 110deg " but got "0 1 0 - 40deg "
FAIL Compositing: property <rotate> underlying [1 2 3 40deg] from add [2 4 6 10deg] to add [3 6 9 50deg] at (-1) should be [1 2 3 10deg] assert_equals: expected "1 2 3 10deg " but got "1 2 3 40deg "
FAIL Compositing: property <rotate> underlying [1 2 3 40deg] from add [2 4 6 10deg] to add [3 6 9 50deg] at (0) should be [1 2 3 50deg] assert_equals: expected "1 2 3 50deg " but got "1 2 3 40deg "
FAIL Compositing: property <rotate> underlying [1 2 3 40deg] from add [2 4 6 10deg] to add [3 6 9 50deg] at (0.25) should be [1 2 3 60deg] assert_equals: expected "1 2 3 60deg " but got "1 2 3 40deg "
FAIL Compositing: property <rotate> underlying [1 2 3 40deg] from add [2 4 6 10deg] to add [3 6 9 50deg] at (0.75) should be [1 2 3 80deg] assert_equals: expected "1 2 3 80deg " but got "1 2 3 40deg "
FAIL Compositing: property <rotate> underlying [1 2 3 40deg] from add [2 4 6 10deg] to add [3 6 9 50deg] at (1) should be [1 2 3 90deg] assert_equals: expected "1 2 3 90deg " but got "1 2 3 40deg "
FAIL Compositing: property <rotate> underlying [1 2 3 40deg] from add [2 4 6 10deg] to add [3 6 9 50deg] at (2) should be [1 2 3 130deg] assert_equals: expected "1 2 3 130deg " but got "1 2 3 40deg "
FAIL Compositing: property <rotate> underlying [1 2 3 270deg] from add [1 2 3 90deg] to replace [0 1 0 100deg] at (-1) should be [-5.49276e-17 -1 -1.64783e-16 100deg] assert_equals: expected "0 - 1 0 100deg " but got "1 2 3 270deg "
FAIL Compositing: property <rotate> underlying [1 2 3 270deg] from add [1 2 3 90deg] to replace [0 1 0 100deg] at (0) should be [1 2 3 360deg] assert_equals: expected "1 2 3 360deg " but got "1 2 3 270deg "
FAIL Compositing: property <rotate> underlying [1 2 3 270deg] from add [1 2 3 90deg] to replace [0 1 0 100deg] at (0.25) should be [-1.20172e-16 1 -3.60516e-16 25deg] assert_equals: expected "0 1 0 25deg " but got "1 2 3 270deg "
FAIL Compositing: property <rotate> underlying [1 2 3 270deg] from add [1 2 3 90deg] to replace [0 1 0 100deg] at (0.75) should be [-1.51909e-17 1 -4.55726e-17 75deg] assert_equals: expected "0 1 0 75deg " but got "1 2 3 270deg "
FAIL Compositing: property <rotate> underlying [1 2 3 270deg] from add [1 2 3 90deg] to replace [0 1 0 100deg] at (1) should be [0 1 0 100deg] assert_equals: expected "0 1 0 100deg " but got "1 2 3 270deg "
FAIL Compositing: property <rotate> underlying [1 2 3 270deg] from add [1 2 3 90deg] to replace [0 1 0 100deg] at (2) should be [-3.3235e-17 -1 -9.97049e-17 160deg] assert_equals: expected "0 - 1 0 160deg " but got "1 2 3 270deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [0 1 0 100deg] at (-1) should be [-5.49276e-17 -1 -1.64783e-16 100deg] assert_equals: expected "0 - 1 0 100deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [0 1 0 100deg] at (0) should be [1 2 3 360deg] assert_equals: expected "1 2 3 360deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [0 1 0 100deg] at (0.25) should be [-1.20172e-16 1 -3.60516e-16 25deg] assert_equals: expected "0 1 0 25deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [0 1 0 100deg] at (0.75) should be [-1.51909e-17 1 -4.55726e-17 75deg] assert_equals: expected "0 1 0 75deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [0 1 0 100deg] at (1) should be [0 1 0 100deg] assert_equals: expected "0 1 0 100deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [0 1 0 100deg] at (2) should be [-3.3235e-17 -1 -9.97049e-17 160deg] assert_equals: expected "0 - 1 0 160deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 0 0 90deg] from add [0 1 0 180deg] to replace [0 0 1 90deg] at (-1) should be [-6.12323e-17 -1 1.57009e-16 90deg] assert_equals: expected "0 - 1 0 90deg " but got "1 0 0 90deg "
FAIL Compositing: property <rotate> underlying [1 0 0 90deg] from add [0 1 0 180deg] to replace [0 0 1 90deg] at (0) should be [-4.32978e-17 -0.707107 -0.707107 180deg] assert_equals: expected "0 - 0.71 - 0.71 180deg " but got "1 0 0 90deg "
FAIL Compositing: property <rotate> underlying [1 0 0 90deg] from add [0 1 0 180deg] to replace [0 0 1 90deg] at (0.25) should be [-1.48952e-16 -0.894427 -0.447214 131.81deg] assert_equals: expected "0 - 0.89 - 0.45 131.81deg " but got "1 0 0 90deg "
FAIL Compositing: property <rotate> underlying [1 0 0 90deg] from add [0 1 0 180deg] to replace [0 0 1 90deg] at (0.75) should be [-2.94392e-17 -0.707107 0.707107 70.5288deg] assert_equals: expected "0 - 0.71 0.71 70.53deg " but got "1 0 0 90deg "
FAIL Compositing: property <rotate> underlying [1 0 0 90deg] from add [0 1 0 180deg] to replace [0 0 1 90deg] at (1) should be [90deg] assert_equals: expected "90deg " but got "1 0 0 90deg "
FAIL Compositing: property <rotate> underlying [1 0 0 90deg] from add [0 1 0 180deg] to replace [0 0 1 90deg] at (2) should be [-6.12323e-17 -1 -4.71028e-16 90deg] assert_equals: expected "0 - 1 0 90deg " but got "1 0 0 90deg "
FAIL Compositing: property <rotate> underlying [none] from add [none] to replace [0 1 0 100deg] at (-1) should be [0 1 0 -100deg] assert_equals: expected "0 1 0 - 100deg " but got "none "
PASS Compositing: property <rotate> underlying [none] from add [none] to replace [0 1 0 100deg] at (0) should be [none]
FAIL Compositing: property <rotate> underlying [none] from add [none] to replace [0 1 0 100deg] at (0.25) should be [0 1 0 25deg] assert_equals: expected "0 1 0 25deg " but got "none "
FAIL Compositing: property <rotate> underlying [none] from add [none] to replace [0 1 0 100deg] at (0.75) should be [0 1 0 75deg] assert_equals: expected "0 1 0 75deg " but got "none "
FAIL Compositing: property <rotate> underlying [none] from add [none] to replace [0 1 0 100deg] at (1) should be [0 1 0 100deg] assert_equals: expected "0 1 0 100deg " but got "none "
FAIL Compositing: property <rotate> underlying [none] from add [none] to replace [0 1 0 100deg] at (2) should be [0 1 0 200deg] assert_equals: expected "0 1 0 200deg " but got "none "
FAIL Compositing: property <rotate> underlying [none] from add [2 4 6 270deg] to replace [none] at (-1) should be [2 4 6 540deg] assert_equals: expected "2 4 6 540deg " but got "none "
FAIL Compositing: property <rotate> underlying [none] from add [2 4 6 270deg] to replace [none] at (0) should be [2 4 6 270deg] assert_equals: expected "2 4 6 270deg " but got "none "
FAIL Compositing: property <rotate> underlying [none] from add [2 4 6 270deg] to replace [none] at (0.25) should be [2 4 6 202.5deg] assert_equals: expected "2 4 6 202.5deg " but got "none "
FAIL Compositing: property <rotate> underlying [none] from add [2 4 6 270deg] to replace [none] at (0.75) should be [2 4 6 67.5deg] assert_equals: expected "2 4 6 67.5deg " but got "none "
PASS Compositing: property <rotate> underlying [none] from add [2 4 6 270deg] to replace [none] at (1) should be [none]
FAIL Compositing: property <rotate> underlying [none] from add [2 4 6 270deg] to replace [none] at (2) should be [2 4 6 -270deg] assert_equals: expected "2 4 6 - 270deg " but got "none "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [none] to replace [0 1 0 100deg] at (-1) should be [0.31 -0.22 0.92 131.66deg] assert_equals: expected "0.31 - 0.22 0.92 131.66deg " but got "1 2 3 90deg "
PASS Compositing: property <rotate> underlying [1 2 3 90deg] from add [none] to replace [0 1 0 100deg] at (0) should be [1 2 3 90deg]
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [none] to replace [0 1 0 100deg] at (0.25) should be [0.21 0.73 0.64 86.72deg] assert_equals: expected "0.21 0.73 0.64 86.72deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [none] to replace [0 1 0 100deg] at (0.75) should be [0.07 0.97 0.21 92.05deg] assert_equals: expected "0.07 0.97 0.21 92.05deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [none] to replace [0 1 0 100deg] at (1) should be [0 1 0 100deg] assert_equals: expected "0 1 0 100deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [none] to replace [0 1 0 100deg] at (2) should be [-0.2 0.79 -0.59 151.11deg] assert_equals: expected "- 0.2 0.79 - 0.59 151.11deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [none] at (-1) should be [1 2 3 720deg] assert_equals: expected "1 2 3 720deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [none] at (0) should be [1 2 3 360deg] assert_equals: expected "1 2 3 360deg " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [none] at (0.25) should be [1 2 3 270deg] assert_equals: expected "1 2 3 270deg " but got "1 2 3 90deg "
PASS Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [none] at (0.75) should be [1 2 3 90deg]
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [none] at (1) should be [none] assert_equals: expected "none " but got "1 2 3 90deg "
FAIL Compositing: property <rotate> underlying [1 2 3 90deg] from add [2 4 6 270deg] to replace [none] at (2) should be [1 2 3 -360deg] assert_equals: expected "1 2 3 - 360deg " but got "1 2 3 90deg "
